+++ tests/client/paging.py
@@ -1,5 +1,6 @@
 import pytest
 
+from tekore import BadRequest
 from ._resources import album_id